requirements.txt
python-dotenv==0.20.0
requests==2.28.1
tqdm==4.64.0
tweepy==4.10.0
